+++ src/main/java/com/CuteNekoDragon/Core/common/CommonProxy.java
@@ -4,6 +4,8 @@ import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import com.CuteNekoDragon.Core.SVOCore;
+import com.CuteNekoDragon.Core.common.data.SVOCreativeTab;
+import com.CuteNekoDragon.Core.common.data.SVOItems;
 import com.CuteNekoDragon.Core.config.SVOConfig;
 
 public class CommonProxy {
@@ -15,5 +17,8 @@ public class CommonProxy {
         bus.addListener(SVOConfig::onLoad);
 
         SVOCore.REGISTRATE.registerEventListeners(bus);
+
+        SVOItems.init();
+        SVOCreativeTab.init();
     }
 }